# Conflicts:
#	app/scripts/controllers/SearchCtrl.js

+++ paw-webapp/webapp/src/main/java/ar/edu/itba/paw/webapp/controller/UserJerseyController.java
@@ -1,17 +1,10 @@
 package ar.edu.itba.paw.webapp.controller;
 
-import ar.edu.itba.paw.webapp.dto.UserDto;
-import ar.edu.itba.paw.webapp.dto.UserGameScoreDto;
-import ar.edu.itba.paw.webapp.dto.UserGameStatusDto;
+import ar.edu.itba.paw.webapp.dto.*;
 import ar.edu.itba.paw.webapp.exceptions.IllegalParameterValueException;
 import ar.edu.itba.paw.webapp.exceptions.MissingJsonException;
-import ar.edu.itba.paw.webapp.interfaces.SessionService;
-import ar.edu.itba.paw.webapp.interfaces.SortDirection;
-import ar.edu.itba.paw.webapp.interfaces.UserDao;
-import ar.edu.itba.paw.webapp.interfaces.UserService;
-import ar.edu.itba.paw.webapp.model.Authority;
-import ar.edu.itba.paw.webapp.model.Game;
-import ar.edu.itba.paw.webapp.model.User;
+import ar.edu.itba.paw.webapp.interfaces.*;
+import ar.edu.itba.paw.webapp.model.*;
 import org.apache.commons.io.IOUtils;
 import org.apache.tika.exception.TikaException;
 import org.apache.tika.metadata.Metadata;
@@ -21,6 +14,7 @@ import org.apache.tika.parser.ParseContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.DefaultHandler;
@@ -29,10 +23,8 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.*;
 import java.io.*;
 import java.net.URI;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
+import java.util.stream.Collectors;
 
 import static ar.edu.itba.paw.webapp.controller.UserJerseyController.END_POINT;
 
@@ -46,16 +38,24 @@ import static ar.edu.itba.paw.webapp.controller.UserJerseyController.END_POINT;
 public class UserJerseyController implements UpdateParamsChecker {
 
     public static final String END_POINT = "users";
-
     @Autowired
-    private UserJerseyController(UserService userService, SessionService sessionService) {
+    private UserJerseyController(UserService userService, SessionService sessionService, MailService mailService, PasswordEncoder passwordEncoder, ShelfService shelfService) {
         this.userService = userService;
         this.sessionService = sessionService;
+        this.mailService = mailService;
+        this.passwordEncoder = passwordEncoder;
+        this.shelfService = shelfService;
     }
 
     private final UserService userService;
 
+    private final MailService mailService;
+
+    private final ShelfService shelfService;
+
     private final SessionService sessionService;
+
+    private PasswordEncoder passwordEncoder;
 
     @Context
     private UriInfo uriInfo;
@@ -139,11 +139,23 @@ public class UserJerseyController implements UpdateParamsChecker {
 
 
     @PUT
-    @Path("/{id : \\d+}/password")
+    @Path("/{id : \\d+}/password-change")
     public Response changePassword(@PathParam("id") final long userId,
                                    final UserDto userDto) {
         checkUpdateValues(userId, "id", userDto);
-        userService.changePassword(userId, userDto.getPassword(), userId); // TODO: updater
+        String newPassword = passwordEncoder.encode(userDto.getPassword());
+        userService.changePassword(userId, newPassword, userId); // TODO: updater
+        mailService.sendEmailChangePassword(userService.findById(userId));
+        return Response.noContent().build();
+    }
+
+    @POST
+    @Path("/{id : \\d+}/password-reset")
+    public Response resetPassword(@PathParam("id") final long userId) {
+        String newPassword = userService.generateNewPassword();
+        String hashedPassword = passwordEncoder.encode(newPassword);
+        userService.changePassword(userId, hashedPassword, userId); // TODO: updater
+//        mailService.sendEmailResetPassword(userService.findById(userId), newPassword);
         return Response.noContent().build();
     }
 
@@ -154,8 +166,9 @@ public class UserJerseyController implements UpdateParamsChecker {
     // ==== Play status ====
 
     @GET
-    @Path("/{id : \\d+}/play-statuses")
-    public Response getPlayStatuses(@PathParam("id") final long userId,
+    @Path("/{id : \\d+}/play-status/{gameId : \\d+}")
+    public Response getPlayStatus(@PathParam("id") final long userId,
+                                  @PathParam("gameId") Long gameIdFilter,
                                     // Pagination and Sorting
                                     @QueryParam("orderBy") @DefaultValue("game-id")
                                     final UserDao.PlayStatusAndGameScoresSortingType sortingType,
@@ -163,8 +176,6 @@ public class UserJerseyController implements UpdateParamsChecker {
                                     final SortDirection sortDirection,
                                     @QueryParam("pageSize") @DefaultValue("25") final int pageSize,
                                     @QueryParam("pageNumber") @DefaultValue("1") final int pageNumber,
-                                    // Filters
-                                    @QueryParam("gameId") Long gameIdFilter,
                                     @QueryParam("gameName") String gameNameFilter) {
         if (userId <= 0) {
             throw new IllegalParameterValueException("id");
@@ -181,8 +192,18 @@ public class UserJerseyController implements UpdateParamsChecker {
 
     }
 
+    @GET
+    @Path("/play-statuses")
+    public Response getPlayStatuses() {
+        return Response.ok(new GenericEntity<List<PlayStatus>>(playStatusList()){}).build();
+    }
+
+    private List<PlayStatus> playStatusList() {
+        return Arrays.stream(PlayStatus.values()).collect(Collectors.toList());
+    }
+
     @POST
-    @Path("/{id : \\d+}/play-statuses")
+    @Path("/{id : \\d+}/play-status")
     @Consumes(value = {MediaType.APPLICATION_JSON})
     public Response addPlayStatus(@PathParam("id") final long userId,
                                   final UserGameStatusDto userGameStatusDto) {
@@ -193,7 +214,7 @@ public class UserJerseyController implements UpdateParamsChecker {
     }
 
     @DELETE
-    @Path("/{id : \\d+}/play-statuses/{gameId : \\d+}")
+    @Path("/{id : \\d+}/play-status/{gameId : \\d+}")
     public Response removePlayStatus(@PathParam("id") final long userId,
                                      @PathParam("gameId") final long gameId) {
         if (userId <= 0) {
@@ -202,7 +223,8 @@ public class UserJerseyController implements UpdateParamsChecker {
         if (gameId <= 0) {
             throw new IllegalParameterValueException("gameId");
         }
-        userService.removePlayStatus(userId, gameId, userId); // TODO: updater
+        userService.setPlayStatus(userId, gameId, PlayStatus.NO_PLAY_STATUS, userId); // TODO: updater
+        if(!belongsToGameList(userId, gameId)) userService.removePlayStatus(userId, gameId, userId);
         return Response.noContent().build();
     }
 
@@ -246,6 +268,9 @@ public class UserJerseyController implements UpdateParamsChecker {
         checkUpdateValues(userId, "id", userGameScoreDto);
         userService.setGameScore(userId, userGameScoreDto.getGameId(), userGameScoreDto.getScore(), userId); // TODO: updater
         final URI uri = uriInfo.getAbsolutePathBuilder().path(String.valueOf(userGameScoreDto.getGameId())).build();
+        if(userService.getPlayStatuses(userId, userGameScoreDto.getGameId(), null, 1, 1, UserDao.PlayStatusAndGameScoresSortingType.GAME_ID, SortDirection.ASC).getData().isEmpty()){
+            userService.setPlayStatus(userId, userGameScoreDto.getGameId(), PlayStatus.NO_PLAY_STATUS, userId);
+        }
         return Response.created(uri).status(Response.Status.CREATED).build();
     }
 
@@ -260,7 +285,14 @@ public class UserJerseyController implements UpdateParamsChecker {
             throw new IllegalParameterValueException("gameId");
         }
         userService.removeGameScore(userId, gameId, userId); // TODO: updater
+        if(!belongsToGameList(userId, gameId)){
+            deleteFromGameList(userId, gameId);
+        }
         return Response.noContent().build();
+    }
+
+    private void deleteFromGameList(long userId, long gameId) {
+        userService.removePlayStatus(userId, gameId, userId);
     }
 
 
@@ -281,7 +313,7 @@ public class UserJerseyController implements UpdateParamsChecker {
      * Creates a map to be used in the
      * {@link #getGameScores(long, UserDao.PlayStatusAndGameScoresSortingType, SortDirection, int, int, Long, String)}
      * or the
-     * {@link #getPlayStatuses(long, UserDao.PlayStatusAndGameScoresSortingType, SortDirection, int, int, Long, String)}
+     * {@link #getPlayStatus(long,Long, UserDao.PlayStatusAndGameScoresSortingType, SortDirection, int, int, String)}
      * methods.
      *
      * @param gameIdFilter   Filter for game id.
@@ -342,6 +374,36 @@ public class UserJerseyController implements UpdateParamsChecker {
                 .header("Access-Control-Allow-Headers", "Content-Type");  //Required by CORS
         return result.build();
     }
+
+    @OPTIONS
+    @Path("/{id : \\d+}/play-status")
+    public Response playStatusOptions() {
+        Response.ResponseBuilder result = Response
+                .ok()
+                .type(MediaType.TEXT_HTML)                                              //Required by CORS
+                .header("Access-Control-Allow-Methods", "PUT,DELETE,OPTIONS,POST,GET")
+                .header("Access-Control-Allow-Headers", "Content-Type");  //Required by CORS
+        return result.build();
+    }
+
+    @OPTIONS
+    @Path("/{id : \\d+}/game-scores")
+    public Response scoreOptions() {
+        return playStatusOptions();
+    }
+
+    @OPTIONS
+    @Path("/{id : \\d+}/play-status/{gameId : \\d+}")
+    public Response playStatusOptions2() {
+        return playStatusOptions();
+    }
+
+    @OPTIONS
+    @Path("/{id : \\d+}/game-scores/{gameId : \\d+}")
+    public Response scoreOptions2() {
+        return playStatusOptions();
+    }
+
 
     @PUT
     @Path("/picture")
@@ -457,5 +519,59 @@ public class UserJerseyController implements UpdateParamsChecker {
 
         String mimeType = metadata.get(HttpHeaders.CONTENT_TYPE);
         return mimeType;
+    }
+
+    @GET
+    @Path("/{id : \\d+}/recommended-games")
+    public Response getRecommendedGames(@PathParam("id") final long userId,
+                                        @QueryParam("shelves") final List<String> shelves) {
+        if (userId <= 0) {
+            throw new IllegalParameterValueException("id");
+        }
+        Collection<Game> recommendedGames = userService.recommendGames(userId, shelves);
+        return Response.ok(new GenericEntity<List<GameDto>>(GameDto.createList(recommendedGames)) {
+        }).build();
+    }
+
+    @GET
+    @Path("/{userId : \\d+}/game-list")
+    public Response listShelfGames(@PathParam("userId") final long userId,
+                                   @QueryParam("orderBy") @DefaultValue("game-id")
+                                   final UserDao.PlayStatusAndGameScoresSortingType sortingType,
+                                   @QueryParam("sortDirection") @DefaultValue("ASC") final SortDirection sortDirection,
+                                   @QueryParam("pageSize") @DefaultValue("25") final int pageSize,
+                                   @QueryParam("pageNumber") @DefaultValue("1") final int pageNumber) {
+
+        JerseyControllerHelper.checkParameters(JerseyControllerHelper
+                .getPaginationReadyParametersWrapper(pageSize, pageNumber)
+                .addParameter("userId", userId, id -> id <= 0));
+
+        return JerseyControllerHelper
+                .createCollectionGetResponse(
+                        uriInfo, sortingType.toString().toLowerCase(), sortDirection,
+                        //TODO Add sorting and filtering to gameList
+                        userService.getGameList(userId, pageNumber, pageSize, sortingType, sortDirection),
+                        (gamesPage) -> new GenericEntity<List<UserGameStatusDto>>(UserGameStatusDto
+                                .createList(gamesPage.getData())) {
+                        },
+                        scoreAndStatusMap(null,null));
+    }
+    
+    /**
+     *
+     * @return whether or not the game belongs to the User's GameList.
+     */
+
+    private boolean belongsToGameList(final long userId, final long gameId){
+        boolean hasPlayStatus = false;
+        UserGameStatus ugs = userService.getPlayStatuses(userId, gameId, null, 1, 1, UserDao.PlayStatusAndGameScoresSortingType.GAME_ID,SortDirection.ASC).getData().iterator().next();
+        if(ugs == null ) {
+            userService.setPlayStatus(userId, gameId, PlayStatus.NO_PLAY_STATUS, userId);
+        } else {
+            if( !ugs.getPlayStatus().equals(PlayStatus.NO_PLAY_STATUS)) hasPlayStatus = true;
+        }
+        return !userService.getGameScores(userId, gameId, null, 1, 1, UserDao.PlayStatusAndGameScoresSortingType.GAME_ID,SortDirection.ASC).getData().isEmpty()
+                || !shelfService.getUserShelves(userId, null, gameId, null, 1, 1, ShelfDao.SortingType.ID, SortDirection.ASC).isEmpty()
+                || hasPlayStatus;
     }
 }
